// insis-kit/src/app/model/plateforme.ts
export interface Plateforme {
    id_plateforme: number
    nom_1_plateforme: string
    nom_2_plateforme: string
    url_img_acceuil_plateforme: string
    nom_general_nat_plateforme: string
    desc_generale_nat_plateforme: string
    desc_page_eqt_plateforme: string
    desc_page_real_plateforme: string
    url_img_page_real_plateforme: string
    titre_page_real_plateforme: string
    desc_page_modes_plateforme: string
    desc_page_contact_plateforme: string
    id_site: number
    id_reseau: number
    titre_page_acceuil_plateforme : string
    desc_page_acceuil_plateforme : string
}